fix: SCRUM-30 소셜 로그인 상단 ui 수정
